Added switch to get back to old behaviour

+++ test/test_datahandling.py
@@ -330,7 +330,7 @@ class TestField(unittest.TestCase):
         spectrum_reference[65] = np.sqrt(2*np.pi)
         spectrum_reference[69] = 2*np.sqrt(2*np.pi)
 
-        x = np.linspace(0, 2*np.pi, 128, endpoint=False)
+        x = np.linspace(0, 2*np.pi, 128, endpoint=False, dtype=np.complex128)
         y = ne.evaluate('exp(1.j*x) + 2.*exp(5.j*x)')
         y = dh.Field(y, axes=[dh.Axis(grid=x)])
         yf = y.fft()
@@ -338,7 +338,7 @@ class TestField(unittest.TestCase):
         # this is passed already with the old code
         self.assertAllClose(spectrum_reference, yf, atol=1e-10)
 
-        x = np.linspace(0, 2*np.pi, 128, endpoint=False) + 3*np.pi/4
+        x = np.linspace(0, 2*np.pi, 128, endpoint=False, dtype=np.complex128) + 3*np.pi/4
         y = ne.evaluate('exp(1.j*x) + 2.*exp(5.j*x)')
         y = dh.Field(y, axes=[dh.Axis(grid=x)])
         yf = y.fft()
